+++ lib/tabulatr/generators/tabulatr/table_generator.rb
@@ -23,12 +23,25 @@
 
 module Tabulatr
   module Generators
-    class InstallGenerator < Rails::Generators::Base
-      desc "initialize tabulatr"
+    class TableGenerator < Rails::Generators::NamedBase
+      desc "create TabulatrData Class"
       source_root File.expand_path('../templates', __FILE__)
+      check_class_collision suffix: 'TabulatrData'
 
-      def copy_translation_file
-        copy_file "tabulatr.yml", "config/locales/tabulatr.yml"
+      argument :attributes, type: :array, default: [], banner: 'field:type field:type'
+
+      def create_tabulatr_data_file
+        template 'tabulatr_data.rb', File.join('app/tabulatr_data/', class_path, "#{file_name}_tabulatr_data.rb")
+      end
+
+      private
+
+      def attributes_names
+        [:id] + attributes.select { |attr| !attr.reference? }.map { |a| a.name.to_sym }
+      end
+
+      def association_names
+        attributes.select { |attr| attr.reference? }.map { |a| a.name.to_sym }
       end
     end
   end